+++ pm/CHANGELOG.md
@@ -19,6 +19,11 @@ description: A detailed, chronological record of all changes made to a project, 
 - Fixed infinite loading on generate button in character builder
 - Fixed SQL Constraint error on character deletion
 
+### Enhancements
+
+- Improved character builder prompt
+- Improved categories example prompt
+
 ## [0.1.0-9] - 2025-05-25
 
 ### Fixed